+++ src/ArduinoHA.h
@@ -7,7 +7,6 @@
 #include "device-types/HABinarySensor.h"
 #include "device-types/HAHVAC.h"
 #include "device-types/HASensor.h"
-#include "device-types/HASensor.cpp"
 #include "device-types/HASwitch.h"
 #include "device-types/HATagScanner.h"
 #include "device-types/HATriggers.h"