+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## 1.2.0
 
+**Breaking changes:**
+* Refactored HASensor implementation. Please take a look at updated example in `eamples/sensor/sensor.ino`
+
 **New features:**
 * Added support for HVAC
 * Added support for excluding devices types from the compilation using defines (see `src/ArduinoHADefines.h`)
 * Added support for setting icon in HASwitch and HASensor
 * Added support for setting retain flag in HASwitch
-
+* Added support for text (const char*) payload in HASensor
 
 **Updates:**
 * Optimized codebase and logic in all devices types